lexiangla-scraper/requirements.txt
requests>=2.31.0
markdownify>=0.13.1
python-dotenv>=1.0.0
playwright>=1.40.0
Pillow>=10.0.0
